adding solidworks support
Early version. May need to recompile the CTL file to automatically change M3 S12000 to M3 P100 and to not include the top header.
For now, this will be done manually. ( A user that will find the ressources here to change the postprocessor of solidworks should be able to sucessfully change 10 Gcole lines in a .cnc file without issues )